requirements.txt
django
djangorestframework
psycopg2-binary
gunicorn
python-dotenv
dj-database-url
django-cors-headers